Move 2.0 into olympus-bg
