--- olympus-bg/TODO FOR 2.0.md
@@ -1,11 +1,0 @@
-# TODO for 2.0
-
-- Fevga doMove
-- Remove clamp() from isMoveValid and doMove
-- Tests
-
-# TODO for 2.1
-
-- Fevga special conditions
-- Tavli event listener
-- Bot (move to olympus-bg)
