--- olympus-bg/src/2.0/index.ts
@@ -1,2 +1,0 @@
-// export Tavli
-// export types
