+++ olympus-bg/src/index.ts
@@ -1,7 +1,2 @@
-import { Variant, Player, rollDie } from "./util.js";
-import { Board } from "./game.js";
-import * as portes from "./variants/portes.js";
-import * as plakoto from "./variants/plakoto.js";
-import * as fevga from "./variants/fevga.js";
-
-export { Variant, Player, rollDie, Board, portes, plakoto, fevga };
+// export Tavli
+// export types
