--- olympus-bg/src/variants/plakoto.ts
@@ -1,124 +1,0 @@
-import { Board as GenericBoard } from "../game.js";
-import { range, Pip, Player, clamp, pipDistance, TODO_DELETE_THIS_isTurnPlayer } from "../util.js";
-
-const Plakoto = () => ({
-    // Inherit from generic board
-    ...GenericBoard(),
-
-    // Initialize the board for a game of plakoto
-    initGame() {
-        this.pips[24] = Pip(15, Player.black); // Black moves towards pip 1 (decreasing)
-        this.pips[1] = Pip(15, Player.white); // White moves towards pip 24 (increasing)
-    },
-
-    // Is the move valid?
-    // from:    Move from pip # <eg. 1>
-    // to:      Move to pip # <eg. 4>
-    // return:  Returns a boolean
-    isMoveValid(from: number, to: number): boolean {
-        to = clamp(to);
-        if (this.pips[from].top !== this.turn) return false;
-
-        // Bearing off
-        if (to === 25 || to === 0) {
-            if (this.turn === Player.white && from < 19) return false;
-            if (this.turn === Player.black && from > 6) return false;
-            // Range of all pips excluding the current player's home quadrant
-            const nonHomePips = this.turn === Player.white ? range(1, 18) : range(7, 24);
-            for (const i of nonHomePips) {
-                if (this.pips[i].top === this.turn || this.pips[i].bot === this.turn) return false;
-            }
-            // If bearing off from an non-exact number of pips
-            if (!this.dice.includes(pipDistance(from, to))) {
-                // Check if there's a big enough dice
-                if (this.dice[0] > pipDistance(from, to) || this.dice[1] > pipDistance(from, to)) {
-                    // Range of pips in the player's home quadrant that are further away than the pip they are trying to bear off of
-                    const farHomePips =
-                        this.turn === Player.white ? range(19, from - 1) : range(from + 1, 6);
-                    for (const i of farHomePips) {
-                        if (this.pips[i].top === this.turn || this.pips[i].bot === this.turn)
-                            return false;
-                    }
-                } else {
-                    return false;
-                }
-            }
-        }
-        // Standard move
-        else {
-            if (from < 1 || from > 24 || to < 1 || to > 24) return false;
-            if (this.pips[to].top !== this.turn && this.pips[to].size > 1) return false;
-            if (!this.dice.includes(pipDistance(from, to))) return false;
-        }
-
-        return true;
-    },
-
-    doMove(from: number, to: number) {
-        TODO_DELETE_THIS_isTurnPlayer(this.turn);
-
-        to = clamp(to);
-        this.recentMove = { from, to };
-
-        // From pip
-        if (this.pips[from].size === 1) {
-            this.pips[from].top = Player.neither;
-            this.pips[from].bot = Player.neither;
-        } else if (this.pips[from].size === 2 && this.pips[from].top !== this.pips[from].bot) {
-            this.pips[from].top = this.pips[from].bot;
-        }
-        this.pips[from].size--;
-
-        // To pip
-        if (to === 0 || to === 25) {
-            // Bearing off
-            if (this.turn === Player.white) this.off[Player.white]++;
-            if (this.turn === Player.black) this.off[Player.black]++;
-        } else {
-            if (this.pips[to].size === 0) {
-                this.pips[to].bot = this.turn;
-            }
-            this.pips[to].top = this.turn;
-            this.pips[to].size++;
-        }
-
-        // Handle dice. NOTE: this will only work for 2 distinct values or 4 identical values
-        if (this.dice[0] >= pipDistance(from, to)) this.dice.shift();
-        else this.dice.pop();
-    },
-
-    // Is the board in a state where the game has just ended?
-    // Returns the number of points won
-    isGameOver(): number {
-        TODO_DELETE_THIS_isTurnPlayer(this.turn);
-
-        const home = { [Player.white]: this.pips[1], [Player.black]: this.pips[24] };
-
-        // Both player's starting checkers have been trapped: game is a draw
-        if (this.pips[1].top !== this.pips[1].bot && this.pips[24].top !== this.pips[24].bot) {
-            this.winner = Player.neither;
-            this.turn = Player.neither;
-            return 1;
-        }
-        // Player has borne off all of their checkers
-        else if (this.off[this.turn] === 15) {
-            this.winner = this.turn;
-            this.turn = Player.neither;
-            // if the other player has born off 0 checkers, return 2 points
-            return this.off[this.otherPlayer(this.winner)] === 0 ? 2 : 1;
-        }
-        // If other player's starting checker has been pinned and current player's is safe
-        else if (
-            home[this.otherPlayer()].top !== home[this.otherPlayer()].bot &&
-            home[this.turn].bot !== this.turn
-        ) {
-            this.winner = this.turn;
-            this.turn = Player.neither;
-            return 2;
-        }
-
-        return 0;
-    },
-});
-
-export const Board = Plakoto;
